Add pathMap preference for mapping local to remote paths.
The user can set a preference in brackets.json that maps project roots
to remote paths, e.g.:

  "php-debugger.pathMap": {
    "/Users/admin/Documents/myProject/": "/var/www/myProject/"
  }

Now, php-debugger can properly map breakpoints to a remote Xdebug
instance, and can open local files while stepping through code. If the
file is not available locally (because it is outside the mapping defined
in the user preference), a note will be logged to the php-debugger
console.